- Add new nuodb version 2.1. - Support nuodb web console and rest API (kakawait/nuodb-with-ecosystem will become obsolete)

+++ 2.0/run.sh
@@ -9,7 +9,7 @@ DEFAULT_ROUTE=$(/sbin/ip route | awk '/default/ { print $3 }')
 export BROKER_ALT_ADDR=${BROKER_ALT_ADDR:-$DEFAULT_ROUTE}
 
 start_nuoagent() {
-    sudo -u nuodb /bin/bash -c "SHELL=/bin/bash java -jar /opt/nuodb/jar/nuoagent.jar --broker > /dev/null 2>&1 &"
+    sudo -u nuodb /bin/bash -c "SHELL=/bin/bash NUODB_USER=nuodb java -jar /opt/nuodb/jar/nuoagent.jar --broker > /dev/null 2>&1 &"
     STATUS=1
     while [[ STATUS -ne 0 ]]; do
         echo "=> Waiting for nuoagent service startup"
@@ -22,7 +22,9 @@ start_nuoagent() {
 
 stop_nuoagent() {
     PID=$(ps -ef | grep "nuoagent" | grep -v grep| awk '{print $2}')
-    kill -9 $PID &>/dev/null
+    if [ -n "$PID" ]; then
+        kill -9 $PID &>/dev/null
+    fi
     echo "=> Nuoagent stopped"
 }
 
@@ -39,9 +41,10 @@ fi
 if [ -f /nuodb-override/default.properties ]; then
     cp /nuodb-override/default.properties /opt/nuodb/etc/default.properties.tpl
 fi
-rm -f /opt/nuodb/etc/default.properties
+rm -f /opt/nuodb/etc/default.properties 
 envsubst < "/opt/nuodb/etc/default.properties.tpl" > "/opt/nuodb/etc/default.properties" 
 chown nuodb:nuodb /opt/nuodb/etc/default.properties
+envsubst < '/etc/supervisor/conf.d/supervisord.conf.tpl' > '/etc/supervisor/conf.d/supervisord.conf'; 
 
 start_nuoagent
 
@@ -55,7 +58,6 @@ fi
 touch /opt/nuodb/data/$DATABASE_NAME/.init
 echo -e "\t=> Create te process from database $DATABASE_NAME"
 /opt/nuodb/bin/nuodbmgr --broker localhost --user $DOMAIN_USER --password ${DOMAIN_PASSWORD:-$RAND_DOMAIN_PASSWORD} --command "start process te host $BROKER_ALT_ADDR database $DATABASE_NAME options '--dba-user $DBA_USER --dba-password ${DBA_PASSWORD:-$RAND_DBA_PASSWORD} --verbose info,warn,error'" &>/dev/null
-    
 if [ -n "$LICENSE" ]; then
     echo -e "\t=> Install nuodb license"
     echo $LICENSE > /license.file
@@ -82,7 +84,16 @@ while IFS= read -r line
 do
     echo  "    $line"
 done <<< "$SUMMARY"
+echo ""
+echo "    ecosystem:"
+if [ "$NUORESTSVC" = true ]; then
+    echo "    [autoconsole] localhost:8888"
+    echo "    [autoconsole-admin] localhost:8889"
+fi
+if [ "$NUOWEBCONSOLE" = true ]; then
+    echo "    [web console] localhost:8080"
+fi
 echo "=========================================================================================="
 echo ""
 
-sudo -u nuodb /bin/bash -c "SHELL=/bin/bash java -jar /opt/nuodb/jar/nuoagent.jar"
+supervisord -n -e $LOG_LEVEL -c /etc/supervisor/conf.d/supervisord.conf
